move control_flow tests into analyze_test.rs
